Move 3d gazers into submodule and rename files appropriately
