+++ pupil_src/shared_modules/gaze_mapping/__init__.py
@@ -12,8 +12,7 @@ from . import matching
 from . import gazer_base
 from .gazer_base import CalibrationError
 from .gazer_2d import Gazer2D
-from .gazer_3d_v1x import Gazer3D_v1x
-from .gazer_hmd_v1x import GazerHMD3D_v1x
+from .gazer_3d import Gazer3D_v1x, GazerHMD3D_v1x
 
 
 def registered_gazer_classes() -> list:
