+++ pupil_src/shared_modules/gaze_mapping/gazer_3d/bundle_adjustment.py
