+++ pupil_src/shared_modules/gaze_mapping/utils.py
@@ -159,7 +159,6 @@ def closest_matches_monocular_batch(ref_pts, pupil, max_dispersion=1 / 15.0):
     return matched
 
 
-
 def _find_nearest_idx(array, value):
     """Find the index of the element in array which is closest to value"""
 
@@ -171,13 +170,3 @@ def _find_nearest_idx(array, value):
             return idx
     except IndexError:
         return idx - 1
-
-
-def _clamp_norm_point(pos):
-    """realistic numbers for norm pos should be in this range.
-        Grossly bigger or smaller numbers are results bad exrapolation
-        and can cause overflow erorr when denormalized and cast as int32.
-    """
-    return min(100.0, max(-100.0, pos[0])), min(100.0, max(-100.0, pos[1]))
-
-
